requirements.txt
canonicalwebteam.flask-base==1.1.0
canonicalwebteam.image-template==1.3.1
google-api-python-client==2.70.0
django-openid-auth==0.17
Flask-OpenID==1.3.0
beautifulsoup4==4.11.1
python-dateutil==2.8.2
lxml==4.9.2
cachetools==5.3.0
black==23.1.0
flake8==6.0.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
webargs==8.2.0
